feat: 후원하는 경우 다른 메세지 나오도록 구현
