+++ src/components/Participate.js
@@ -1,8 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import { Close } from "@styled-icons/evaicons-solid/Close";
 import { withRouter } from "react-router-dom";
-
+import { StyledClose, SubmitBtn } from "components/sharedComponents";
 const Wrapper = styled.section`
   width: 100%;
   height: 100vh;
@@ -34,28 +33,6 @@ const LabelWrapper = styled.div`
   }
 `;
 
-const SubmitBtn = styled.div`
-  width: 100%;
-  background: ${(props) => props.theme.color.blue};
-  height: 50px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-weight: bold;
-  color: #fff;
-  border-radius: 10px;
-  font-size: ${(props) => props.theme.size.mmd};
-`;
-
-const StyledClose = styled(Close)`
-  position: absolute;
-  top: 10px;
-  right: 20px;
-  width: 30px;
-  height: 30px;
-  color: black;
-`;
-
 const Alert = styled.div`
   font-size: ${(props) => props.theme.size.smd};
   color: red;
@@ -63,70 +40,9 @@ const Alert = styled.div`
   height: 20px;
 `;
 
-const ModalContainer = styled.section`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  display: ${(props) => (props.visible ? "block" : "none")};
-`;
-
-const ModalOverlay = styled.div`
-  background: rgba(0, 0, 0, 0.4);
-  width: 100%;
-  height: 100%;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 15px;
-`;
-
-const ModalContent = styled.div`
-  position: relative;
-  background: #fff;
-  border: 1px solid black;
-  border-radius: 10px;
-  width: 100%;
-  min-height: 150px;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  text-align: left;
-  line-height: 1.3em;
-
-  & > * {
-    margin: 5px 0;
-  }
-  & > div:nth-child(2) {
-    font-size: ${(props) => props.theme.size.smd};
-  }
-`;
-
-const ModalHeader = styled.div`
-  font-size: ${(props) => props.theme.size.mmd};
-  font-weight: bold;
-`;
 // 후원 참석자 폼 컴포넌트
 const Participate = React.memo(
-  ({
-    nickname,
-    phoneNumber,
-    comment,
-    amount,
-    onChange,
-    onAddComment,
-    alert,
-    goBack,
-    visible,
-    onToggleModal,
-  }) => {
+  ({ form, onChange, alert, goBack, onToggleModal }) => {
     return (
       <>
         <Wrapper>
@@ -138,7 +54,7 @@ const Participate = React.memo(
               name="nickname"
               id="nickname-input"
               onChange={onChange}
-              value={nickname}
+              value={form.nickname}
               placeholder="닉네임"
             />
             {alert.nickname && (
@@ -152,7 +68,7 @@ const Participate = React.memo(
               name="phoneNumber"
               id="phoneNumber-input"
               onChange={onChange}
-              value={phoneNumber}
+              value={form.phoneNumber}
               placeholder="010-0000-0000"
             />
             {alert.phoneNumber && (
@@ -168,7 +84,7 @@ const Participate = React.memo(
               rows="5"
               maxLength="100"
               onChange={onChange}
-              value={comment}
+              value={form.comment}
               placeholder="응원합니다"
             ></textarea>
           </LabelWrapper>
@@ -178,7 +94,7 @@ const Participate = React.memo(
               name="amount"
               id="amount-select"
               onChange={onChange}
-              value={amount}
+              value={form.amount}
             >
               <option value="0">0</option>
               <option value="5000">5000</option>
@@ -189,17 +105,6 @@ const Participate = React.memo(
           <SubmitBtn onClick={onToggleModal}>응원 참여하기</SubmitBtn>
           <StyledClose onClick={goBack} />
         </Wrapper>
-        <ModalContainer visible={visible}>
-          <ModalOverlay>
-            <ModalContent>
-              <StyledClose onClick={() => onAddComment(1)} />
-              <ModalHeader>아직 구현되지 않은 기능입니다.</ModalHeader>
-              <div>
-                기부 기능이 오픈되면 가장 먼저 연락 메세지를 드리겠습니다.
-              </div>
-            </ModalContent>
-          </ModalOverlay>
-        </ModalContainer>
       </>
     );
   }
